Allow importing (not using) in non-browser envs
This allows the module to be imported (or required) in a non-browser
(read: nodejs) environment, but throws when getCaretCoordinates is called.

This is especially useful for the React Components use case. Typically a
React Component is defined in a file which includes both render-time and
after-mounting-time dependencies, among which textarea-caret-position.
This patch allows the React Component definition file to be correctly
executed for the purpose of testing or server-side rendering, although it
(correctly) throws when attempting to call getCaretCoordinates.

This commit also fixes minor quoting inconcistencies.